+++ plugin-server/src/capabilities.ts
@@ -22,6 +22,7 @@ export function getPluginServerCapabilities(config: PluginsServerConfig): Plugin
                 cdpInternalEvents: true,
                 cdpCyclotronWorker: true,
                 cdpCyclotronWorkerPlugins: true,
+                cdpCyclotronWorkerFetch: true,
                 cdpApi: true,
             }
 
@@ -72,6 +73,10 @@ export function getPluginServerCapabilities(config: PluginsServerConfig): Plugin
         case PluginServerMode.cdp_cyclotron_worker_plugins:
             return {
                 cdpCyclotronWorkerPlugins: true,
+            }
+        case PluginServerMode.cdp_cyclotron_worker_fetch:
+            return {
+                cdpCyclotronWorkerFetch: true,
             }
         case PluginServerMode.cdp_api:
             return {
